fix(develop): update folder structure
