+++ pubspec.yaml
@@ -36,9 +36,11 @@ dependencies:
   flutter_riverpod: ^2.3.2
 
 dev_dependencies:
+  build_runner: ^2.3.3
   flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
+  freezed: ^2.3.2
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
